+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # develop
   * Feature: Asterisk calls now support the Join API (joining only)
+  * Bugfix: The Asterisk translator now ignores calls to 'h' or of type 'Kill'
 
 # v0.9.2 - 2012-02-18
   * Feature: Asterisk calls receiving media commands are implicitly answered